Move Talk.js
