+++ app/scripts.babel/class/Talk.js
