+++ app/scripts.babel/main.js
@@ -1,6 +1,6 @@
 import 'babel-polyfill';
 import HasValue from './class/HasValue';
-import Talk from './Talk';
+import Talk from './class/Talk';
 import Message from './Message';
 import TalkArea from './TalkArea';
 import formatDate from './formatDate';
